Move state scripts to separate package.
Use separate `statescript` package for all related implementation.

Changelog: None
